Use the world's random instead of using the global one
When generating you shouldn't use the global random module but use the
random object provided in the world to prevent other worlds from messing
with your randomness and potentially makes parts of the generation
unreproducible. The API is the exact same between the random module and
world.random so there's no drawback to using that.